refactor(studio): drain remaining commands from main.rs into modules
Move the last command groups out of the Tauri entry point into themed
modules under commands/, leaving main.rs as a true entry point (state
wrapper, OSC send helpers, setup + generate_handler! registry):

- app.rs      get_state, OSC config load/save, About box, auto-tune snapshots
- gain.rs     speaker/object/master gain, mute, loudness, auto-gain
- diag.rs     metering/diag rates, diag toggle, gain-table subscribe handshake
- engine.rs   config save/reload, log level, ramp mode, DRC, layout export
- layout_io.rs layout select/import/export + native file-dialog pickers
- render.rs   (appended) distance-diffuse + render bridge_path/input_pipe

No behavior change: command names, signatures and the generate_handler!
registry are identical, so the OSC addresses and JS invoke contract are
unchanged. main.rs: 836 -> 301 lines.
