+++ omniphony-studio/src-tauri/src/commands/engine.rs
@@ -1,0 +1,106 @@
+//! Engine-level controls: renderer config save/reload, log level, ramp mode,
+//! dynamic-range-control (DRC) tuning and the layout-export trigger.
+//!
+//! Each command forwards a value to the renderer over OSC.
+
+use crate::osc_listener::OscControlMsg;
+use crate::{send_control, SharedState};
+use tauri::State;
+
+#[tauri::command]
+pub fn control_save_config(state: State<SharedState>) {
+    send_control(
+        &state.osc_tx,
+        OscControlMsg::SendNoArgs {
+            address: "/omniphony/control/save_config".to_string(),
+        },
+    );
+}
+
+#[tauri::command]
+pub fn control_reload_config(state: State<SharedState>) {
+    send_control(
+        &state.osc_tx,
+        OscControlMsg::SendNoArgs {
+            address: "/omniphony/control/reload_config".to_string(),
+        },
+    );
+}
+
+#[tauri::command]
+pub fn control_log_level(state: State<SharedState>, value: String) {
+    let trimmed = value.trim().to_ascii_lowercase();
+    if !matches!(
+        trimmed.as_str(),
+        "off" | "error" | "warn" | "info" | "debug" | "trace"
+    ) {
+        return;
+    }
+    send_control(
+        &state.osc_tx,
+        OscControlMsg::SendString {
+            address: "/omniphony/control/log_level".to_string(),
+            value: trimmed,
+        },
+    );
+}
+
+#[tauri::command]
+pub fn control_ramp_mode(state: State<SharedState>, value: String) {
+    let trimmed = value.trim().to_ascii_lowercase();
+    if !matches!(trimmed.as_str(), "off" | "frame" | "sample") {
+        return;
+    }
+    send_control(
+        &state.osc_tx,
+        OscControlMsg::SendString {
+            address: "/omniphony/control/ramp_mode".to_string(),
+            value: trimmed,
+        },
+    );
+}
+
+#[tauri::command]
+pub fn control_drc_mode(state: State<SharedState>, value: String) {
+    send_control(
+        &state.osc_tx,
+        OscControlMsg::SendString {
+            address: "/omniphony/control/input/drc_mode".to_string(),
+            value,
+        },
+    );
+}
+
+#[tauri::command]
+pub fn control_drc_weight(state: State<SharedState>, value: f32) {
+    send_control(
+        &state.osc_tx,
+        OscControlMsg::SendFloat {
+            address: "/omniphony/control/input/drc_weight".to_string(),
+            value: value.clamp(0.0, 1.0),
+        },
+    );
+}
+
+#[tauri::command]
+pub fn control_export_layout(state: State<SharedState>, name: Option<String>) {
+    if let Some(raw) = name {
+        let trimmed = raw.trim();
+        if !trimmed.is_empty() {
+            send_control(
+                &state.osc_tx,
+                OscControlMsg::SendString {
+                    address: "/omniphony/control/layout/export".to_string(),
+                    value: trimmed.to_string(),
+                },
+            );
+            return;
+        }
+    }
+    send_control(
+        &state.osc_tx,
+        OscControlMsg::SendNoArgs {
+            address: "/omniphony/control/layout/export".to_string(),
+        },
+    );
+}
