+++ omniphony-studio/src-tauri/src/main.rs
@@ -9,21 +9,24 @@ mod osc_listener;
 mod osc_parser;
 
 use std::path::PathBuf;
-use std::sync::atomic::{AtomicI32, Ordering};
+use std::sync::atomic::AtomicI32;
 use std::sync::{Arc, Mutex};
 
 use app_state::AppState;
-use config::{load_config, save_config, OscConfig};
-use layouts::Layout;
+use config::load_config;
 use osc_listener::{spawn_osc_task, OscControlMsg};
-use rfd::FileDialog;
 use tauri::{Manager, State};
 use tokio::sync::mpsc::UnboundedSender;
 
 // Tauri command handlers grouped into themed modules under `commands/`. Glob-
 // imported so `generate_handler!` can keep referring to them by bare name.
+use commands::app::*;
 use commands::audio::*;
+use commands::diag::*;
+use commands::engine::*;
+use commands::gain::*;
 use commands::input::*;
+use commands::layout_io::*;
 use commands::mpv_overlay::*;
 use commands::orender::*;
 use commands::render::*;
@@ -79,544 +82,6 @@ pub(crate) fn send_distance_metric(state: &State<SharedState>, address: &str, va
         OscControlMsg::SendString {
             address: address.to_string(),
             value: normalized,
-        },
-    );
-}
-
-#[derive(serde::Serialize)]
-struct AboutInfo {
-    name: &'static str,
-    version: &'static str,
-    license: &'static str,
-    repository_url: &'static str,
-    description: &'static str,
-}
-
-// ── Tauri commands ────────────────────────────────────────────────────────
-
-#[tauri::command]
-fn get_state(state: State<SharedState>) -> serde_json::Value {
-    let s = state.inner.lock().unwrap();
-    serde_json::to_value(&*s).unwrap_or(serde_json::Value::Null)
-}
-
-#[tauri::command]
-fn get_osc_config(state: State<SharedState>) -> OscConfig {
-    load_config(&state.config_dir)
-}
-
-#[tauri::command]
-fn get_about_info() -> AboutInfo {
-    AboutInfo {
-        name: "Omniphony Studio",
-        version: env!("CARGO_PKG_VERSION"),
-        license: "GPL-3.0-only",
-        repository_url: "https://github.com/mgth/Omniphony",
-        description: "Omniphony is an open spatial-audio project built around realtime rendering, transport, control, and monitoring tools for object-based audio workflows. Omniphony Studio is the visual control surface of that ecosystem.",
-    }
-}
-
-#[tauri::command]
-fn save_osc_config(state: State<SharedState>, config: OscConfig) -> Result<(), String> {
-    save_config(&state.config_dir, &config)?;
-    state.inner.lock().unwrap().osc_metering_enabled =
-        Some(if config.osc_metering_enabled { 1 } else { 0 });
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SetMeteringEnabled {
-            enabled: config.osc_metering_enabled,
-        },
-    );
-    let listen_port = *state.listen_port.lock().unwrap();
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::Reconnect {
-            host: config.host,
-            rx_port: config.osc_rx_port,
-            listen_port,
-        },
-    );
-    Ok(())
-}
-
-#[tauri::command]
-fn control_osc_metering(state: State<SharedState>, enable: i32) -> Result<(), String> {
-    let enabled = enable != 0;
-    let mut cfg = load_config(&state.config_dir);
-    cfg.osc_metering_enabled = enabled;
-    save_config(&state.config_dir, &cfg)?;
-    state.inner.lock().unwrap().osc_metering_enabled = Some(if enabled { 1 } else { 0 });
-    send_control(&state.osc_tx, OscControlMsg::SetMeteringEnabled { enabled });
-    Ok(())
-}
-
-#[tauri::command]
-fn select_layout(state: State<SharedState>, key: String) -> bool {
-    let mut s = state.inner.lock().unwrap();
-    let exists = s.layouts.iter().any(|l| l.key == key);
-    if exists {
-        s.selected_layout_key = Some(key);
-    }
-    exists
-}
-
-#[tauri::command]
-fn import_layout_from_path(
-    state: State<SharedState>,
-    path: String,
-) -> Result<serde_json::Value, String> {
-    let trimmed = path.trim();
-    if trimmed.is_empty() {
-        return Err("empty layout path".to_string());
-    }
-    let mut layout = layouts::load_layout_file(std::path::Path::new(trimmed))
-        .ok_or_else(|| "failed to parse layout file".to_string())?;
-
-    let mut s = state.inner.lock().unwrap();
-    let base_key = layout.key.clone();
-    let mut suffix = 1usize;
-    while s.layouts.iter().any(|l| l.key == layout.key) {
-        layout.key = format!("{base_key}-{}", suffix);
-        suffix += 1;
-    }
-    s.selected_layout_key = Some(layout.key.clone());
-    s.layouts.push(layout);
-    s.layouts
-        .sort_by(|a, b| a.name.to_lowercase().cmp(&b.name.to_lowercase()));
-
-    Ok(serde_json::json!({
-        "layouts": s.layouts,
-        "selectedLayoutKey": s.selected_layout_key
-    }))
-}
-
-#[tauri::command]
-fn pick_import_layout_path() -> Option<String> {
-    FileDialog::new()
-        .add_filter("Layout", &["json", "yaml", "yml"])
-        .pick_file()
-        .map(|path| path.to_string_lossy().to_string())
-}
-
-#[tauri::command]
-fn pick_export_layout_path(suggested_name: Option<String>) -> Option<String> {
-    let file_name = suggested_name
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(|s| {
-            let lowered = s.to_ascii_lowercase();
-            if lowered.ends_with(".yaml") || lowered.ends_with(".yml") || lowered.ends_with(".json")
-            {
-                s.to_string()
-            } else {
-                format!("{s}.yaml")
-            }
-        })
-        .unwrap_or_else(|| "layout.yaml".to_string());
-
-    FileDialog::new()
-        .add_filter("Layout YAML", &["yaml", "yml"])
-        .add_filter("Layout JSON", &["json"])
-        .set_file_name(&file_name)
-        .save_file()
-        .map(|path| path.to_string_lossy().to_string())
-}
-
-#[tauri::command]
-fn pick_import_evaluation_artifact_path() -> Option<String> {
-    FileDialog::new()
-        .add_filter("Omniphony evaluator", &["oevl"])
-        .pick_file()
-        .map(|path| path.to_string_lossy().to_string())
-}
-
-#[tauri::command]
-fn pick_export_evaluation_artifact_path(suggested_name: Option<String>) -> Option<String> {
-    let file_name = suggested_name
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(|s| {
-            let lowered = s.to_ascii_lowercase();
-            if lowered.ends_with(".oevl") {
-                s.to_string()
-            } else {
-                format!("{s}.oevl")
-            }
-        })
-        .unwrap_or_else(|| "evaluation.oevl".to_string());
-
-    FileDialog::new()
-        .add_filter("Omniphony evaluator", &["oevl"])
-        .set_file_name(&file_name)
-        .save_file()
-        .map(|path| path.to_string_lossy().to_string())
-}
-
-#[tauri::command]
-fn pick_bridge_path() -> Option<String> {
-    FileDialog::new()
-        .set_title("Select bridge library")
-        .pick_file()
-        .map(|path| path.to_string_lossy().to_string())
-}
-
-#[tauri::command]
-fn pick_orender_path() -> Option<String> {
-    FileDialog::new()
-        .set_title("Select orender executable")
-        .pick_file()
-        .map(|path| path.to_string_lossy().to_string())
-}
-
-#[tauri::command]
-fn export_layout_to_path(path: String, layout: Layout) -> Result<(), String> {
-    let trimmed = path.trim();
-    if trimmed.is_empty() {
-        return Err("empty export path".to_string());
-    }
-
-    layouts::save_layout_file(std::path::Path::new(trimmed), &layout)
-}
-
-#[tauri::command]
-fn control_speaker_gain(state: State<SharedState>, id: i32, gain: f32) {
-    let clamped = gain.max(0.0).min(2.0);
-    let seq = state.realtime_seq.fetch_add(1, Ordering::Relaxed) + 1;
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendArgs {
-            address: "/omniphony/control/realtime/speaker_gain".to_string(),
-            args: vec![
-                rosc::OscType::Int(id),
-                rosc::OscType::Float(clamped),
-                rosc::OscType::Int(seq),
-            ],
-        },
-    );
-}
-
-#[tauri::command]
-fn control_object_mute(state: State<SharedState>, id: i32, muted: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: format!("/omniphony/control/object/{id}/mute"),
-            value: if muted != 0 { 1 } else { 0 },
-        },
-    );
-}
-
-#[tauri::command]
-fn control_object_gain(state: State<SharedState>, id: String, gain: f32) {
-    let clamped = gain.clamp(0.0, 2.0);
-    let seq = state.realtime_seq.fetch_add(1, Ordering::Relaxed) + 1;
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendArgs {
-            address: "/omniphony/control/realtime/object_gain".to_string(),
-            args: vec![
-                rosc::OscType::String(id),
-                rosc::OscType::Float(clamped),
-                rosc::OscType::Int(seq),
-            ],
-        },
-    );
-}
-
-#[tauri::command]
-fn control_speaker_mute(state: State<SharedState>, id: i32, muted: i32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/speakers",
-        serde_json::json!({
-            "speakerEdits": [{
-                "id": id.max(0),
-                "muted": muted != 0
-            }]
-        }),
-    );
-}
-
-#[tauri::command]
-fn control_master_gain(state: State<SharedState>, gain: f32) {
-    let clamped = gain.max(0.0).min(2.0);
-    let seq = state.realtime_seq.fetch_add(1, Ordering::Relaxed) + 1;
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendArgs {
-            address: "/omniphony/control/realtime/master_gain".to_string(),
-            args: vec![rosc::OscType::Float(clamped), rosc::OscType::Int(seq)],
-        },
-    );
-}
-
-#[tauri::command]
-fn control_loudness(state: State<SharedState>, enable: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: "/omniphony/control/loudness".to_string(),
-            value: if enable != 0 { 1 } else { 0 },
-        },
-    );
-}
-
-#[tauri::command]
-fn control_auto_gain(state: State<SharedState>, enable: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: "/omniphony/control/auto_gain".to_string(),
-            value: if enable != 0 { 1 } else { 0 },
-        },
-    );
-}
-
-#[tauri::command]
-fn control_auto_gain_ceiling(state: State<SharedState>, db: f32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/auto_gain_ceiling".to_string(),
-            value: db.clamp(-12.0, 0.0),
-        },
-    );
-}
-
-#[tauri::command]
-fn auto_tune_snapshot_save(state: State<SharedState>, snapshot: serde_json::Value) {
-    *state.auto_tune_snapshot.lock().unwrap() = Some(snapshot);
-}
-
-#[tauri::command]
-fn auto_tune_snapshot_take(state: State<SharedState>) -> Option<serde_json::Value> {
-    state.auto_tune_snapshot.lock().unwrap().take()
-}
-
-#[tauri::command]
-fn auto_tune_snapshot_peek(state: State<SharedState>) -> Option<serde_json::Value> {
-    state.auto_tune_snapshot.lock().unwrap().clone()
-}
-
-#[tauri::command]
-fn control_metering_rate_hz(state: State<SharedState>, value: f32) {
-    let clamped = value.max(1.0).min(1000.0);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/metering/rate_hz".to_string(),
-            value: clamped,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_diag_rate_hz(state: State<SharedState>, value: f32) {
-    let clamped = value.max(1.0).min(1000.0);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/diag/rate_hz".to_string(),
-            value: clamped,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_diag_publication_enabled(state: State<SharedState>, enable: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: "/omniphony/control/diag/enabled".to_string(),
-            value: if enable != 0 { 1 } else { 0 },
-        },
-    );
-}
-
-/// Subscribe to one speaker's per-band gain field. `have_version` is the version
-/// already cached on this client (0 if none); `speaker_index` is the speaker to
-/// display. The renderer pushes that speaker's field only if the version differs,
-/// then on every topology rebuild while subscribed. Sent on first consumer, on
-/// speaker change, and as a 5 s heartbeat (idempotent, self-healing).
-#[tauri::command]
-fn subscribe_speaker_gaintable(state: State<SharedState>, have_version: i32, speaker_index: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendArgs {
-            address: "/omniphony/control/debug/speaker_gaintable/subscribe".to_string(),
-            args: vec![
-                rosc::OscType::Int(have_version.max(0)),
-                rosc::OscType::Int(speaker_index.max(0)),
-            ],
-        },
-    );
-}
-
-/// Unsubscribe from the gain-table push stream (last consumer released). The
-/// client keeps its cached table; a later re-subscribe negotiates by version.
-#[tauri::command]
-fn unsubscribe_speaker_gaintable(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/debug/speaker_gaintable/unsubscribe".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_distance_diffuse_enabled(state: State<SharedState>, enable: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: "/omniphony/control/distance_diffuse/enabled".to_string(),
-            value: if enable != 0 { 1 } else { 0 },
-        },
-    );
-}
-
-#[tauri::command]
-fn control_distance_diffuse_threshold(state: State<SharedState>, value: f32) {
-    let v = value.max(0.01);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/distance_diffuse/threshold".to_string(),
-            value: v,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_distance_diffuse_curve(state: State<SharedState>, value: f32) {
-    let v = value.max(0.0);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/distance_diffuse/curve".to_string(),
-            value: v,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_save_config(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/save_config".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_reload_config(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/reload_config".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_log_level(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(
-        trimmed.as_str(),
-        "off" | "error" | "warn" | "info" | "debug" | "trace"
-    ) {
-        return;
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/log_level".to_string(),
-            value: trimmed,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_ramp_mode(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(trimmed.as_str(), "off" | "frame" | "sample") {
-        return;
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/ramp_mode".to_string(),
-            value: trimmed,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_drc_mode(state: State<SharedState>, value: String) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/drc_mode".to_string(),
-            value,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_drc_weight(state: State<SharedState>, value: f32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/input/drc_weight".to_string(),
-            value: value.clamp(0.0, 1.0),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_export_layout(state: State<SharedState>, name: Option<String>) {
-    if let Some(raw) = name {
-        let trimmed = raw.trim();
-        if !trimmed.is_empty() {
-            send_control(
-                &state.osc_tx,
-                OscControlMsg::SendString {
-                    address: "/omniphony/control/layout/export".to_string(),
-                    value: trimmed.to_string(),
-                },
-            );
-            return;
-        }
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/layout/export".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_render_bridge_path(state: State<SharedState>, value: String) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/render/bridge_path".to_string(),
-            value: value.trim().to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_render_input_pipe(state: State<SharedState>, value: String) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/render/input_pipe".to_string(),
-            value: value.trim().to_string(),
         },
     );
 }
